Scroller: Implement internal scroll reducer
This allows configuring the number of steps required for a  high resolution scroll event
by adjusting the SCROLLER_CONFIG.internal_divisor value. A value of 120 is the default for
low resolution scrolling. If the device is set to high resolution scrolling by a
get_report request the scroll multiplier will be set to 1. In this state, 1 step will
result in one high resolution scroll event. Increasing the value, n, of internal_divisor
will result in n steps being required before a scroll event occurs.

Signed-off-by: Cameron Ewing <cameronramsayewing@gmail.com>