Scroller: Handle USB suspend/resume
Update the USB status callback to properly support the device being
suspended and resumed by the host. A semaphore, usb_suspend_sem, was
added to indicate state. The device is not suspended if the semaphore's
value is 1. On suspend the callback takes the semaphore, on device reset
and resume the semaphore is given.

Additionally the behavior of usb_conf_sem is more explicitly defined in
the status callback. On USB connection, disconnection, and reset the
configuration semaphore is taken. Only on USB configuration is the
semaphore given.

The semaphore values are now checked in send_hid_report() to validate
device state before attempting to send a report. The device must be
configured and not suspended to send a report.

Signed-off-by: Cameron Ewing <cameronramsayewing@gmail.com>