Scroller: Improve validation of feature reports
scroller_hog.c: Validate the data recieved by a write_feature_report
call. Checks to see if the recieved type will fit in a 2 byte buffer and
sends an invalid offset error if it does not.

Additionally fixed the warning on pointer arithmatic associated with
the memcpy indexing by casting to a byte size.

Signed-off-by: Cameron Ewing <cameronramsayewing@gmail.com>