Scroller: Detect pairing button press
Use the input subsystem to detect long presses of a button for using as
a pairing button.

prj.conf: Enable Input and Input Longpress

nrf52840dk_nrf52840.overlay: Add compatible for longpress events
attached to button 1 on the development kit. Emit KEY_A for short and
KEY_X for longpress. Interval for longpress is one second.

scroller_ble.c: Create input callback and attach it to the subsystem.

Signed-off-by: Cameron Ewing <cameronramsayewing@gmail.com>