Scroller: Use workqueue for starting advertising
Rather than directly starting advertising in the main thread, move it to
a workqueue job so that the scheduler can run it when there is time.
This prevents delaying state changes in main.

main.c: Submit the request to advertise and check it was submitted
without error.

scroller_ble.c: Refactor scroller_ble_advertise_pairing to
advertise_pairing and define a workqueue handler, advertise_work that
calls it. Move input handling code to end of file. Fix a typo in the
disconnect state change.

scroller_ble.h: Expose advertise_work publicly so the main thread can
use it.

Signed-off-by: Cameron Ewing <cameronramsayewing@gmail.com>