scroller_usb.c: Refactor state processing
Pull usb_state processing to a function and always publish
usb_state_events. Subscribe to the usb_state_events and process state
change based on them.

State processing only occurs on a state transition:
If the device is not already suspended, suspend the device. If it
becomes USB_STATE_CONFIGURED, wake the device.

Signed-off-by: Cameron Ewing <cameronramsayewing@gmail.com>